Tweak encoder sensitivity/resolution.
On the left half, I currently have:

https://splitkb.com/products/linear-rotary-encoder

> This encoder actuates thirty times on a full cycle, so one full twist will
> perform thirty actions.

> The resolution of this encoder in QMK is 4, which is the default setting for
> encoders.

And on the right:

https://splitkb.com/products/industrial-rotary-encoder

> This encoder actuates thirty times on a full cycle, so one full twist will
> perform thirty actions.

> The resolution of this encoder in QMK is 2.

According to
https://github.com/zmkfirmware/zmk/blob/01d2102c2326b86b0f87bb008c2a3eb3871e3963/app/boards/shields/kyria/kyria.dtsi#L61-L77,
the default resolutions in ZMK for the Kyria are set to 4 for both halves, but
I'm only seeing 15 actuations per full cycle with the left encoder, so
hopefully this change fixes that. (Perhaps the units differ between ZMK and
QMK.)

Note that I haven't tried the right encoder yet, as it isn't
functional (https://github.com/zmkfirmware/zmk/pull/728 hasn't landed yet).